embeddings/requirements.txt
functions-framework==3.10.1
google-cloud-storage==3.10.0